Add method to cleanup entries that'll never run
When this package is used in a service, it's useful to be able to remove
jobs that are not required anymore.

Since the package does not provide a way to remove a specific entry,
this change adds a method to remove entries that'll never get run in
future.  With this change Schedule implementations can return a
zero-time from Next(), and such entries can be removed from Cron.