Fix data race on concurrent calls to cron Entries()
The current code uses a channel for Entries() to signal that it wants a
snapshot by writing nil, the run() thread to realize that someone wants
a snapshot, and uses the same channel for Entries() to read it back.

When two threads call Entries() simultaneously, one of them can receive
the others request, which should have gone to the run() thread.

This commit adds a test which reproduces the problem, and fixes it by
making a new channel for requests, so the channels are always
unidirectional.