update cron schedule + test coverage.
Added a function Cron.UpdateSchedule(id, spec) that can be used
to update the schedule of the cron job associated with the given id.
The following changes/additions were made.
1. To allow for constant time lookup, added a new member to Cron.
	entryLookupTable map[EntryID]*Entry

2. Added a new function, addNewEntry(*Entry) that not only appends
the entry to the list of entries but also adds it to the
entryLookupTable.

3. Retrofitted Cron.Schedule() and Cron.run() to now use
Cron.addNewEntry() instead of directly appending Cron.entries.

4. Retrofitted Cron.removeEntry() to also delete the entry from
entryLookupTable.

Added test coverage for cron.UpdateSchedule(id, spec).
The following tests were added.
1. Update the schedule without calling cron.Start() and verify if
the schedule interval changed.
2. Update the schedule while the cron job is running (after calling
cron.Start()) and verify if the schedule interval changed.