reinit timer after sched update + fix data race.
Made sure that the new schedule is forwarded via channels to ensure
that there isn't any race condition.
After updating entry.Schedule and entry.Next, the entries are reordered
as entries[0] might no longer be the next entry to run. With this,
the timer is also reinitialized.
Got rid of entriesLookupTable as it is not necessary for this change.

Added support for passing either a spec string or a Schedule when
updating the schedule of an entry.

Updated test code to test that the schedule indeed was updated and
that the timer was reinitialized.

Got rid of the following dependencies.
1. github.com/stretchr/testify/assert
2. github.com/pkg/errors

Ran "go test" and "go test -race" on the project.